+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [Unreleased]
 
+- Limited keyboard height to 50% of screen
 - Fixed key previews disappearing on change app
 - Fixed key preview vertical position on devices with soft buttons